Renamed folder and namespace network2 to http.
To avoid multiple commits that will rename network2 to network after
porting is completed it has been decided to rename namespace and folder
to _http_. This is a more descriptive name for the actual use of the
Network classes as the main point of http(s) requests handler.

Relates-to: OLPEDGE-368
